removed deprecated setup.py, changed package name to "sompackage"
