+++ .gitignore
@@ -3,6 +3,9 @@ __pycache__/
 *.py[cod]
 *$py.class
 
+# .pypirc file
+*.pypirc
+
 # C extensions
 *.so
