+++ sompackage/visualization/__init__.py
@@ -1,0 +1,1 @@
+from . import dotmap, histogram, hitmap, mapview, umatrix
