--- sompy/visualization/__init__.py
@@ -1,1 +1,0 @@
-from . import dotmap, histogram, hitmap, mapview, umatrix
